+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: `NavBar`: Add `small` value on `modifier` prop.
 
 ## [7.1.0] - 2022-01-20